Adds storage of raw data by timestamp in log files.
Data storage now runs on a cronjob. Currently, the cronjob demands
root permissions, which is nonideal. We would like to ask for a
preferred user argument at initialization.

Additionally, data is now stored with a timestamp. Log files are
stored in subdirectories corresponding to their place in the json
heirarchy. Ideally, we would like to use that heirarchy to build
an arbitrary RESTful-ready JSON payload.