Adds documentation and catches missing json file.
We can now allow json file to be missing in action and new one to be created.
Still need to do the same for log files. Currently, touched blank logs.
The primary reason for not accounting for missing files is that we would
like to restructure to allow aggregation by timestamp.

Cron job is being run by the current user, not root, now. This is
advantageous because we can put the expectation of permissions on
operations and not developers, who can be prone to making mistakes.

We are also okay with the cron job creating a new isntance of
Diagnostics because we are attempting to load from the file before
trying to create a new json dictionary.