Adds revamped version of Diagnostics as MetricsEntry.
The primary issue with Diagnostics was that it was trying to record the
information AND store it. MetricsEntry is simply the data collection
process localized to a single object. MetricsEntry comes with a
dictionary field containing all the percieved relevant metrics and the
timestamp at object instantiation.

Arguably, this timestamp might be slightly off due to time spent
recording metrics, but it shouldn't be long enough to matter.