Adds metric logging to complete diagnostics replacement.
The logging class should make organization of individual metrics
trivial. Some standard output print statements were added for debugging
as well. Files are stored by an hourly naming scheme. Even if the
process starts up in the middle of an hour, the files will all be
uniformly named to match the start of the hour.

We've gotten rid of the individual file management. The rationale was
that as more metrics were added, the overhead time to add features
outweighed the ease of data access. Besides, you can just json.load the
hourly file and be ready to go.