Replaces Cron job with APScheduler library management.
The python service shouldn't use the cron to maintain logging times.
While the cron is a reliable source of regular activity and doesn't
block the process from doing other things, it's a bad idea to play with
the system cron or even a user's cron.

The scheduler class is necessary because we wish to create a new log
file every hour. The apscheduler job scope doesn't allow for new objects
to be instantiated and assigned outside of a class's members.

client.py also ensures that the logs folder is available at runtime and
now is able to delete all past logs and retrieve the current hourly
metrics.