Adds documentation and updated README for revamp.
One big nice thing about the current scheme is that creating aggregators
to be endpoints for REST GET and POST requests is quite easy. The data
is always present starting when the service is first run, and log files
are simply dumped json, so an aggregator simply has to load the
appropriate log files matching a date range and then unpack that data.

One problem with this solution is that in order to access ONLY CPU
statistics, all data must still be unpacked. This could be taxing on
systems with low available virtual memory. It's part of the trade-off of
moving away from individual log files per metric recorded.